+++ backend/src/main/java/com/ktds/questionformentoring/member/service/MemberService.java
@@ -20,5 +20,5 @@ public interface MemberService {
     public List<MemberManagementDTO> searchMemberTypeList(String mbrTypeCd);
 
     // 관리자 페이지 - 유저 삭제
-    public void deleteByMbrNo(Long mbrNo);
+    public void deleteByMbrNo(int mbrNo);
 }
